pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "slices"
version = "0.1.0"
description = "SLICES: Self-Supervised Learning for Intensive Care Embeddings System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0",
    "lightning>=2.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "polars>=0.19",
    "pyarrow>=14.0",
    "duckdb>=0.9",
    "wandb>=0.15",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "rich>=13.0",
    "matplotlib>=3.10.7",
    "seaborn>=0.13.2",
    "pre-commit>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = [
    "N806",  # Variable in function should be lowercase (allow B, T, D for tensor dims)
    "N812",  # Lowercase imported as non-lowercase (allow `import pytorch as L`, `functional as F`)
    "E741",  # Ambiguous variable name (allow `l` for layer)
]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
# Disable errors common with PyTorch/Polars dynamic typing
disable_error_code = [
    "override",
    "assignment",
    "arg-type",
    "return-value",
    "union-attr",
    "call-overload",
    "call-arg",
    "operator",
    "index",
    "attr-defined",
    "misc",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
